% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romigami.R
\name{initialize_environment}
\alias{initialize_environment}
\title{Creates a reticulate virtual environment and install the Python omigami package on it}
\usage{
initialize_environment()
}
\description{
Creates a reticulate virtual environment and install the Python omigami package on it
}
