% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romigami.R
\name{match_spectra_from_path}
\alias{match_spectra_from_path}
\title{Instantiates the Spec2Vec tool of the Omigami client and returns the n_best spectra matches for each input in mgf_path}
\usage{
match_spectra_from_path(token, mgf_path, n_best)
}
\arguments{
\item{token}{Bearer token needed for authentication}

\item{mgf_path}{Path to the mgf file}

\item{n_best}{number of spectra matches to return for each spectrum input}
}
\value{
A list of dataframes result
}
\description{
Instantiates the Spec2Vec tool of the Omigami client and returns the n_best spectra matches for each input in mgf_path
}
