% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romigami.R
\name{omigami_init}
\alias{omigami_init}
\title{Initialize environment and import Omigami module}
\usage{
omigami_init()
}
\value{
Omigami module
}
\description{
Initialize environment and import Omigami module
}
